pyproject.toml
[project]
name = "newsscraper"
version = "0.1.0"
description = "News scraper with sentiment analysis and summarization"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4",
    "requests",
    "textblob",
    "vaderSentiment",
    "newspaper3k",
    "scikit-learn",
    "matplotlib",
    "wordcloud",
    "nltk",
    "lxml",
    "flask",
    "flask-cors",
    "selenium",
    "transformers",
    "pymongo",
    "gunicorn",
    "python-dotenv",
    "celery",
    "redis",
    "fake-useragent",
    "torch",
    "ftfy",
    "black>=24.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
